Move BasicCallableMapping from topology.callable_mapping to topology.mapping
