+++ sympde/topology/callable_mapping.py
@@ -1,51 +1,9 @@
-from abc import ABC, abstractmethod
-
 from sympy import Symbol
 
 from sympde.utilities.utils import lambdify_sympde
-from .mapping import Mapping
+from .mapping import Mapping, BasicCallableMapping
 
-__all__ = ('BasicCallableMapping', 'CallableMapping')
-
-#==============================================================================
-class BasicCallableMapping(ABC):
-    """
-    Transformation of coordinates
-
-    F: R^l -> R^p
-    F(eta) = x
-
-    with l <= p
-    """
-    @abstractmethod
-    def __call__(self, *eta):
-        """ Evaluate mapping at location eta. """
-
-    @abstractmethod
-    def jacobian(self, *eta):
-        """ Compute Jacobian matrix at location eta. """
-
-    @abstractmethod
-    def metric(self, *eta):
-        """ Compute components of metric tensor at location eta. """
-
-    @abstractmethod
-    def metric_det(self, *eta):
-        """ Compute determinant of metric tensor at location eta. """
-
-    @property
-    @abstractmethod
-    def ldim(self):
-        """ Number of logical/parametric dimensions in mapping
-            (= number of eta components).
-        """
-
-    @property
-    @abstractmethod
-    def pdim(self):
-        """ Number of physical dimensions in mapping
-            (= number of x components).
-        """
+__all__ = ('CallableMapping',)
 
 #==============================================================================
 class CallableMapping(BasicCallableMapping):
